[feat] --> move from email to username and subject from token.
