+++ apigateway/src/main/java/vaultweb/apigateway/controller/GatewayAuthController.java
@@ -10,10 +10,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import vaultweb.apigateway.dto.UserDetails;
 import vaultweb.apigateway.dto.request.LoginRequest;
 import vaultweb.apigateway.dto.request.UserRegistrationRequest;
 import vaultweb.apigateway.dto.response.AuthResponse;
+import vaultweb.apigateway.dto.response.UserDetails;
 import vaultweb.apigateway.service.AuthService;
 
 @RestController
