+++ apigateway/src/main/java/vaultweb/apigateway/dto/response/UserDetails.java
@@ -1,0 +1,11 @@
+package vaultweb.apigateway.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record UserDetails(
+        String email,
+        String username,
+        String name
+) {
+}
